reafactor: separate menuBtns functions
